Remove decimal here

+++ curricula/mapping/models.py
@@ -47,7 +47,7 @@ class Problem:
     path: Path
     short: str
     number: int
-    percentage: Decimal
+    percentage: float
 
     title: str
     authors: List[Author]
@@ -68,7 +68,7 @@ class Problem:
 
         authors = list(Author(**author) for author in data.pop("authors"))
         grading = Grading(**data.pop("grading")) if "grading" in data else None
-        percentage = Decimal(reference["percentage"])
+        percentage = reference["percentage"]
         return cls(assignment, path, short, number, percentage, authors=authors, grading=grading, **data)
